refactor: move YarnRoPE class to RoPEUtils.swift and clean up code
